moved exec and execSync back to the original location
