+++ scripts/exec-sync.js
