+++ scripts/exec.js
